Initial Mirror for XWadTools
This is a collection of WAD editing and viewing tools that I think was
originally made available around 2001 and was mainly intended for use on
Unix systems but to be able to compile on Windows too.

I don't know how it was originally distributed or where it was
originally made available.

The README contains some more information, and there's a small entry on
the Doom wiki here:
https://doomwiki.org/wiki/XWadTools