Actions Workflow to Release Tools
This workflow will build the tools on Linux, Windows, and macOS, then create a Github release and
attach the binaries as assets.

The Windows build is performed with the MSYS2 shell, since doing it with the default one (which I'm
guessing is WSL?) was producing Linux binaries. The Windows build agent is supposed to have MSYS2
pre-installed, but that didn't seem to work correctly, so I used some instructions from the MSYS2
website.

To get it to build on Windows, the Makefile was changed to use `MSDOS` as the OS, then after that,
some small modifications were made or removed for conditional compilation. This code is so old that
it was using some functionality from 16 bit MSDOS that wasn't even available on Windows any more.
Some of the other tools wouldn't build on Windows, but the reasons weren't immediately obvious, so I
didn't fix those up.

Perhaps at some point I'll come back and fix up the whole suite.